+++ src/Fenix.Runtime/Common/Message/RegisterClientReq.cs
@@ -14,7 +14,7 @@ namespace Fenix.Common.Message
     public class RegisterClientReq : IMessageWithCallback
     {
         [Key(0)]
-        public UInt32 hostId { get; set; }
+        public UInt64 hostId { get; set; }
 
         [Key(1)]
         public String hostName { get; set; }
